__UI__ Added CustomWindow, a system that add behaviour to UI Window, making everything nice to use Started GameUI (NavMenu, Inventory) Added CustomWindow to the TradeUI Changed HudManager, remove useless Variables Created GameUI.cs which control the GameUI. duh
__PlayerController__
Removed the move of the player controller. Some Tile move system bullshit. Remplaced for the Rigidbody2D

__Art & GameObject__
Fixed the bad resolution of the art, Changed the PixelScale to make them smaller or bigger, and changed back all gameobject scale to 1
Switched filtermode to Point

__Level Design__
Made a new Level01 Map, sorry thom, but yours was shit.

__AudioManager__
Fixed the bug where sometimes the Music wont play or will stop after a loading scene

__LevelManager__
Re-enabled the Fading screen and changed the logic of it to make it better
Calling the hudmanager during the changing scene so the sceneUI will be enable when it should be

__Imported__
Imported UI Texture
Imported Tilemap environement

__Music__
Added music for the Level 1 and the Combat Scene

__Code Cleaning__
Went through some code to clean them because it was causing some stupid bug and I didn't want to waste my time on trying to understand the code itself. The logic and the structure of the code is still there, just removed dead code / useless steps and cleaned the shit out of it.
Example of cleaning: Renaming variables because it was not clear enough. Added Functions, removed some useless if else if else if else if else if else if

__CombatUI__
Added 2 new functions EnemyDead, FriendlyDead to allow the UI To know when a enemy or ally is died so it wont let you select it.
Added Struct, to manage easily the dead stuff ^
Fixed the Index Error  and the EndRound stuff, so now it wont do some random black magic stuff to tell whom turns is it.
*Note to the david (the original dev of the script): Never change the index of a list when you call somewhere else. Enumerator always starts at 0 and end at the Size -1. Keep it like this in the code, that will avoid MULTIPLE errors.
Changed the pre-defined index of the Enemy in the prefab scene -> CombatUI, so the index respect the logic of the programmation, which is, start index 0.

+++ Assets/Scripts/Managers/GameManager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class GameManager : Singleton<GameManager>
 {
     private PlayerController m_Player;
-    private int m_PlayerDistance = 0;
+    private float m_PlayerDistance = 0;
     [SerializeField]
     private int m_Odds = 125;
 
@@ -46,7 +46,7 @@ public class GameManager : Singleton<GameManager>
         LevelManager.Instance.ChangeLevel("DefaultCombatScene", true, 0.5f);
     }
 
-    public void GetPlayerDistance(int aInt)
+    public void GetPlayerDistance(float aInt)
     {
         m_PlayerDistance = aInt;
         System.Random rand = new System.Random();
